+++ src/ApplicationInterface.php
@@ -1,0 +1,19 @@
+<?php
+/*
+*
+* (c) Marco Bunge <marco_bunge@web.de>
+*
+* For the full copyright and license information, please view the LICENSE.txt
+* file that was distributed with this source code.
+*
+* Date: 05.03.2016
+* Time: 00:00
+*/
+
+namespace Proton;
+
+
+interface ApplicationInterface
+{
+
+}